move parse master file
